@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   crossDomain: true,
   target: '#app-wrapper',
   svgSpritePath: 'css/svg/',
@@ -10,21 +10,20 @@ module.exports = {
     l: [1280, 500]
   },
   breakPointsPrefix: 'o-media',
-  defaultControls: [
-    {
-      "name": "scaleline"
-    },
-    {
-      "name": "zoom"
-    },
-    {
-      "name": "rotate"
-    },
-    {
-      "name": "attribution"
-    },
-    {
-      "name": "fullscreen"
-    }
+  defaultControls: [{
+    name: 'scaleline'
+  },
+  {
+    name: 'zoom'
+  },
+  {
+    name: 'rotate'
+  },
+  {
+    name: 'attribution'
+  },
+  {
+    name: 'fullscreen'
+  }
   ]
-}
+};
