@@ -1,29 +1,25 @@
- "use strict";
+import $ from 'jquery';
+import utils from './utils';
 
-var $ = require('jquery');
-var utils = require('./utils');
-
-module.exports = function(target, items, options) {
-  var dataAttribute = 'data-' + options.dataAttribute || 'default';
-  var $target = $('#' + target);
-  var activeItem = options.active || undefined;
-  var activeCls = 'o-active';
-  var ul;
-  var li = [];
-  var cls = 'o-dropdown-li';
-  var icon = utils.createSvg({
+export default function dropDown(target, items, options) {
+  const dataAttribute = `data-${options.dataAttribute}` || 'default';
+  const $target = $(`#${target}`);
+  const activeItem = options.active || undefined;
+  const activeCls = 'o-active';
+  let ul;
+  const li = [];
+  const cls = 'o-dropdown-li';
+  const icon = utils.createSvg({
     href: '#ic_check_24px',
     cls: 'o-icon-24'
   });
-  render();
-  addListener();
 
   function render() {
-    items.forEach(function(item, index) {
-      var obj = {
-        cls: cls
+    items.forEach((item, index) => {
+      const obj = {
+        cls
       };
-      var active = utils.createElement('span', icon, {
+      const active = utils.createElement('span', icon, {
         cls: 'o-icon'
       });
       obj[dataAttribute] = item.value;
@@ -38,11 +34,15 @@ module.exports = function(target, items, options) {
     $target.append(ul);
   }
 
-  function addListener() {
-    $target.on('click', 'ul', function(e) {
-      var $active = $(e.target);
-      var $prevAct = $active.parent().find('.o-active');
+  function toggleActive($active) {
+    $target.find('li').removeClass(activeCls);
+    $active.addClass(activeCls);
+  }
 
+  function addListener() {
+    $target.on('click', 'ul', (e) => {
+      const $active = $(e.target);
+      const $prevAct = $active.parent().find('.o-active');
       $target.trigger({
         type: 'changeDropdown',
         prevSelected: $prevAct.data(options.dataAttribute),
@@ -52,8 +52,6 @@ module.exports = function(target, items, options) {
     });
   }
 
-  function toggleActive($active) {
-    $target.find('li').removeClass(activeCls);
-    $active.addClass(activeCls);
-  }
-};
+  render();
+  addListener();
+}
