@@ -1,8 +1,8 @@
-var getcenter = require('./geometry/getcenter');
-var getarea = require('./geometry/getarea');
+import getcenter from './geometry/getcenter';
+import getarea from './geometry/getarea';
 
-var geom = {};
+const geom = {};
 geom.center = getcenter;
 geom.area = getarea;
 
-module.exports = geom;
+export default geom;
