@@ -1,286 +1,281 @@
-"use strict";
-var ol = require('openlayers');
-var $ = require('jquery');
-var Viewer = require('./viewer');
-var maputils = require('./maputils');
-var getAttributes = require('./getattributes');
-var featureInfo = require('./featureinfo');
+import EsriJSON from 'ol/format/esrijson';
+import Collection from 'ol/collection';
+import $ from 'jquery';
+import viewer from './viewer';
+import maputils from './maputils';
+import getAttributes from './getattributes';
+import featureInfo from './featureinfo';
 
-var map;
+let map;
 
-function getFeaturesFromRemote(evt) {
-        map = Viewer.getMap();
-        var requestResult = [];
-        var requestPromises = getFeatureInfoRequests(evt).map(function(request) {
-            return request.fn.then(function(feature) {
-                var layer = Viewer.getLayer(request.layer);
-
-                if (layer === undefined) {
-                  var groups = Viewer.getLayerGroups();
-                  for (var i = 0; i < groups.length; i++) {
-                    var layers = groups[i].get('layers');
-
-                    var layerInGroup = layers.getArray().filter(function(layer) {
-                      return (layer.get('name') === request.layer)
-                    });
-
-                    if (layerInGroup.length != 0) {
-                      layer = layerInGroup[0];
-                      break;
-                    }
-                  };
-                }
-
-                if(feature) {
-                    requestResult.push({
-                        title: layer.get('title'),
-                        feature: feature,
-                        content: getAttributes(feature, layer)
-                    });
-                    return requestResult;
-                }
-            });
-        });
-        return $.when.apply($, requestPromises).then(function(data) {
-            return requestResult;
-        });
-}
-function getFeatureInfoRequests(evt) {
-    var requests = [];
-    //Check for support of crossOrigin in image, absent in IE 8 and 9
-    if('crossOrigin' in new(Image)) {
-        map.forEachLayerAtPixel(evt.pixel, function(layer) {
-            if (layer.get('queryable')) {
-              var item = getGetFeatureInfoRequest(layer, evt.coordinate);
-              if(item) {requests.push(item)};
-            }
-        });
-    }
-    //If canvas is tainted
-    else if(isTainted(evt.pixel)) {
-        var layers = Viewer.getQueryableLayers();
-        layers.forEach(function(layer) {
-            if (layer.get('queryable')) {
-
-              //If layer is tainted, then create request for layer
-              if(isTainted(evt.pixel, layer)) {
-                  var item = getGetFeatureInfoRequest(layer, evt.coordinate);
-                  if(item) {requests.push(item)};
-              }
-
-              //If layer is not tainted, test if layer hit at pixel
-              else if(layerAtPixel(evt.pixel, layer)){
-                  var item = getGetFeatureInfoRequest(layer, evt.coordinate);
-                  if(item) {requests.push(item)};
-              }
-            }
-        });
-    }
-    //If crossOrigin is not supported and canvas not tainted
-    else {
-        map.forEachLayerAtPixel(evt.pixel, function(layer) {
-            if (layer.get('queryable') === true) {
-              var item = getGetFeatureInfoRequest(layer, evt.coordinate);
-              if (item) {
-                requests.push(item)
-              };
-            }
-        });
-    }
-    return requests;
-}
-function getGetFeatureInfoRequest(layer, coordinate) {
-    var layerType = layer.get('type'),
-    obj = {};
-    obj.layer = layer.get('name');
-
-    switch (layerType) {
-      case 'WMTS':
-        if(layer.get('featureinfoLayer')) {
-            var featureinfoLayerName = layer.get('featureinfoLayer'),
-            featureinfoLayer = Viewer.getLayer(featureinfoLayerName);
-            return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
-        }
-        else {
-            return undefined;
-        }
-        break;
-      case 'WMS':
-        if(layer.get('featureinfoLayer')) {
-            var featureinfoLayerName = layer.get('featureinfoLayer'),
-            featureinfoLayer = Viewer.getLayer(featureinfoLayerName);
-            return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
-        }
-        else {
-            obj.cb = "GEOJSON";
-            obj.fn = getGetFeatureInfoUrl(layer, coordinate);
-            return obj;
-        }
-        case 'AGS_TILE':
-            if(layer.get('featureinfoLayer')) {
-                var featureinfoLayerName = layer.get('featureinfoLayer'),
-                featureinfoLayer = Viewer.getLayer(featureinfoLayerName);
-                return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
-            }
-            else {
-                obj.fn = getAGSIdentifyUrl(layer, coordinate);
-                return obj;
-            }
-      default:
-        return undefined;
-    }
-}
-function isTainted(pixel, layerFilter) {
-    try {
-        if(layerFilter) {
-            map.forEachLayerAtPixel(pixel, function(layer) {
-                return layerFilter === layer;
-            });
-        }
-        else {
-            map.forEachLayerAtPixel(pixel, function(layer) {});
-        }
-        return false;
-    }
-    catch(e) {
-        console.log(e);
-        return true;
-    }
-}
-function layerAtPixel(pixel, matchLayer) {
-    map.forEachLayerAtPixel(pixel, function(layer) {
-        return matchLayer === layer;
-    })
-}
-function getFeaturesAtPixel(evt, clusterFeatureinfoLevel) {
-    map = Viewer.getMap();
-    var result = [],
-    cluster = false;
-    map.forEachFeatureAtPixel(evt.pixel,
-        function(feature, layer) {
-          var l = layer;
-          var queryable = false;
-          if(layer) {
-              queryable = layer.get('queryable');
-          }
-          if(feature.get('features')) {
-              //If cluster
-              var collection = feature.get('features');
-              if (collection.length > 1) {
-                var zoom = map.getView().getZoom();
-                var zoomLimit = clusterFeatureinfoLevel === -1 ? Viewer.getResolutions().length : zoom + clusterFeatureinfoLevel;
-                if(zoomLimit < Viewer.getResolutions().length) {
-                    map.getView().setCenter(evt.coordinate);
-                    map.getView().setZoom(zoom + 1);
-                    cluster = true;
-                    return true;
-                }
-                else {
-                    collection.forEach(function(f) {
-                          var item = {};
-                          item.title = l.get('title');
-                          item.feature = f;
-                          item.content =  getAttributes(f,l);
-                          result.push(item);
-                    });
-                }
-              }
-              else if(collection.length == 1 && queryable) {
-                  var item = {};
-                  item.title = l.get('title');
-                  item.feature = collection[0];
-                  item.content = getAttributes(collection[0],l);
-                  result.push(item);
-              }
-          }
-          else if(queryable) {
-              var item = {};
-              item.title = l.get('title');
-              item.feature = feature;
-              item.content = getAttributes(feature,l)
-              result.push(item);
-          }
-        },
-        {
-          hitTolerance: featureInfo.getHitTolerance()
-        });
-
-    if(cluster) {
-        return false;
-    }
-    else {
-        return result;
-    }
-}
 function getGetFeatureInfoUrl(layer, coordinate) {
-  if (layer.get('queryurl') != undefined) {
-    var url = layer.get('queryurl') + coordinate[1] + ',' + coordinate[0];
+  let url;
+  if (layer.get('queryurl') !== undefined) {
+    url = `${layer.get('queryurl') + coordinate[1]},${coordinate[0]}`;
   } else {
-    var url = layer.getSource().getGetFeatureInfoUrl(
-    coordinate, map.getView().getResolution(), Viewer.getProjection(),
-    {'INFO_FORMAT': 'application/json'});
+    const resolution = map.getView().getResolution();
+    const projection = viewer.getProjection();
+    url = layer.getSource().getGetFeatureInfoUrl(coordinate, resolution, projection, {
+      INFO_FORMAT: 'application/json',
+      FEATURE_COUNT: '20'
+    });
   }
 
-    return $.ajax(url, {
-      type: 'post'
-    })
-    .then(function(response) {
-        if(response.error) {
-            return [];
-        }
-        else if (layer.get('queryurl') != undefined) {
-          var feature = maputils.jsonToPointFeature(response, coordinate);
-          return feature;
-        }
-        else {
-          return maputils.geojsonToFeature(response);
-        }
-    }
-    );
+  return $.ajax(url, {
+    type: 'post'
+  })
+    .then((response) => {
+      if (response.error) {
+        return [];
+      } else if (layer.get('queryurl') !== undefined) {
+        const feature = maputils.jsonToPointFeature(response, coordinate);
+        return feature;
+      }
+      return maputils.geojsonToFeature(response);
+    });
 }
-function getAGSIdentifyUrl(layer, coordinate) {
-  var projectionCode = Viewer.getProjectionCode();
-  var esriSrs = projectionCode.split(':').pop();
-  var layerId = layer.get('id');
-  var source = Viewer.getMapSource()[layer.get('sourceName')];
-  var serverUrl = source.url;
-  var esrijsonFormat = new ol.format.EsriJSON();
-  var size = map.getSize();
-  var tolerance = source.hasOwnProperty('tolerance') ? source.tolerance.toString() : 5;
-  var extent = map.getView().calculateExtent(size);
 
-  var url = serverUrl +
-      '/identify?f=json&' +
-      'returnGeometry=true' +
-      '&geometryType=esriGeometryPoint'+
-      '&sr=' + esriSrs +
-      '&geometry=' + coordinate +
-      '&outFields=*' +
-      '&geometryPrecision=2' +
-      '&tolerance=' + tolerance +
-      '&layers=all:' + layerId +
-      '&mapExtent=' + extent +
-      '&imageDisplay=' + size + ',' + '96';
+function getAGSIdentifyUrl(layer, coordinate) {
+  const projectionCode = viewer.getProjectionCode();
+  const esriSrs = projectionCode.split(':').pop();
+  const layerId = layer.get('id');
+  const source = viewer.getMapSource()[layer.get('sourceName')];
+  const serverUrl = source.url;
+  const esrijsonFormat = new EsriJSON();
+  const size = map.getSize();
+  const tolerance = Object.prototype.hasOwnProperty.call(source, 'tolerance') ? source.tolerance.toString() : 5;
+  const extent = map.getView().calculateExtent(size);
+
+  const url = [`${serverUrl}`,
+    '/identify?f=json',
+    '&returnGeometry=true',
+    '&geometryType=esriGeometryPoint',
+    `&sr=${esriSrs}`,
+    `&geometry=${coordinate}`,
+    '&outFields=*',
+    '&geometryPrecision=2',
+    `&tolerance=${tolerance}`,
+    `&layers=all:${layerId}`,
+    `&mapExtent=${extent}`,
+    `&imageDisplay=${size},96`].join('');
 
   return $.ajax({
-        url: url,
-        dataType: 'jsonp'
+    url,
+    dataType: 'jsonp'
   })
-    .then(function(response) {
-        if(response.error) {
-            return [];
-        }
-        else {
-            var obj = {};
-            obj.features = response.results;
-            var features = esrijsonFormat.readFeatures(obj, {
-                    featureProjection: Viewer.getProjection()
-            });
-            return features[0];
-        }
-    }
-    );
+    .then((response) => {
+      if (response.error) {
+        return [];
+      }
+      const obj = {};
+      obj.features = response.results;
+      const features = esrijsonFormat.readFeatures(obj, {
+        featureProjection: viewer.getProjection()
+      });
+      return features;
+    });
 }
 
-module.exports.getFeaturesFromRemote = getFeaturesFromRemote;
-module.exports.getFeaturesAtPixel = getFeaturesAtPixel;
+function isTainted(pixel, layerFilter) {
+  try {
+    if (layerFilter) {
+      map.forEachLayerAtPixel(pixel, layer => layerFilter === layer);
+    }
+
+    return false;
+  } catch (e) {
+    console.log(e);
+    return true;
+  }
+}
+
+function layerAtPixel(pixel, matchLayer) {
+  map.forEachLayerAtPixel(pixel, layer => matchLayer === layer);
+}
+
+function getGetFeatureInfoRequest(layer, coordinate) {
+  const layerType = layer.get('type');
+  const obj = {};
+  obj.layer = layer.get('name');
+
+  switch (layerType) {
+    case 'WMTS':
+      if (layer.get('featureinfoLayer')) {
+        const featureinfoLayerName = layer.get('featureinfoLayer');
+        const featureinfoLayer = viewer.getLayer(featureinfoLayerName);
+        return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
+      }
+      break;
+    case 'WMS':
+      if (layer.get('featureinfoLayer')) {
+        const featureinfoLayerName = layer.get('featureinfoLayer');
+        const featureinfoLayer = viewer.getLayer(featureinfoLayerName);
+        return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
+      }
+      obj.cb = 'GEOJSON';
+      obj.fn = getGetFeatureInfoUrl(layer, coordinate);
+      return obj;
+    case 'AGS_TILE':
+      if (layer.get('featureinfoLayer')) {
+        const featureinfoLayerName = layer.get('featureinfoLayer');
+        const featureinfoLayer = viewer.getLayer(featureinfoLayerName);
+        return getGetFeatureInfoRequest(featureinfoLayer, coordinate);
+      }
+      obj.fn = getAGSIdentifyUrl(layer, coordinate);
+      return obj;
+    default:
+      return null;
+  }
+
+  return null;
+}
+
+function getFeatureInfoRequests(evt) {
+  const requests = [];
+  // Check for support of crossOrigin in image, absent in IE 8 and 9
+  if ('crossOrigin' in new (Image)()) {
+    map.forEachLayerAtPixel(evt.pixel, (layer) => {
+      if (layer.get('queryable')) {
+        const item = getGetFeatureInfoRequest(layer, evt.coordinate);
+        if (item) {
+          requests.push(item);
+        }
+      }
+    });
+  } else if (isTainted(evt.pixel)) { // If canvas is tainted
+    const layers = viewer.getQueryableLayers();
+    layers.forEach((layer) => {
+      if (layer.get('queryable')) {
+        // If layer is tainted, then create request for layer
+        if (isTainted(evt.pixel, layer)) {
+          const item = getGetFeatureInfoRequest(layer, evt.coordinate);
+          if (item) {
+            requests.push(item);
+          }
+        } else if (layerAtPixel(evt.pixel, layer)) { // If layer is not tainted, test if layer hit at pixel
+          const item = getGetFeatureInfoRequest(layer, evt.coordinate);
+          if (item) {
+            requests.push(item);
+          }
+        }
+      }
+    });
+  } else { // If crossOrigin is not supported and canvas not tainted
+    map.forEachLayerAtPixel(evt.pixel, (layer) => {
+      if (layer.get('queryable') === true) {
+        const item = getGetFeatureInfoRequest(layer, evt.coordinate);
+        if (item) {
+          requests.push(item);
+        }
+      }
+    });
+  }
+  return requests;
+}
+
+function getFeaturesFromRemote(evt) {
+  map = viewer.getMap();
+  const requestResult = [];
+  const requestPromises = getFeatureInfoRequests(evt).map(request => request.fn.then((features) => {
+    let layer = viewer.getLayer(request.layer);
+
+
+    if (layer === undefined) {
+      const groups = viewer.getLayerGroups();
+      for (let i = 0; i < groups.length; i += 1) {
+        const layers = groups[i].get('layers');
+
+        const layerInGroup = layers.getArray().filter(filteredLayer => filteredLayer.get('name') === request.layer);
+
+        if (layerInGroup.length !== 0) {
+          layer = layerInGroup[0];
+          break;
+        }
+      }
+    }
+
+    if (features) {
+      if (features instanceof Collection) {
+        features.forEach((feature) => {
+          requestResult.push({
+            title: layer.get('title'),
+            feature,
+            content: getAttributes(feature, layer)
+          });
+        });
+      } else {
+        requestResult.push({
+          title: layer.get('title'),
+          feature: features,
+          content: getAttributes(features, layer)
+        });
+      }
+      return requestResult;
+    }
+
+    return false;
+  }));
+  return $.when(...requestPromises).then(() => requestResult);
+}
+
+function getFeaturesAtPixel(evt, clusterFeatureinfoLevel) {
+  map = viewer.getMap();
+  const result = [];
+  let cluster = false;
+  map.forEachFeatureAtPixel(evt.pixel, (feature, layer) => {
+    const l = layer;
+    let queryable = false;
+    if (layer) {
+      queryable = layer.get('queryable');
+    }
+    if (feature.get('features') && queryable) {
+      // If cluster
+      const collection = feature.get('features');
+      if (collection.length > 1) {
+        const zoom = map.getView().getZoom();
+        const zoomLimit = clusterFeatureinfoLevel === -1 ? viewer.getResolutions().length : zoom + clusterFeatureinfoLevel;
+        if (zoomLimit < viewer.getResolutions().length) {
+          map.getView().setCenter(evt.coordinate);
+          map.getView().setZoom(zoom + 1);
+          cluster = true;
+          return true;
+        }
+        collection.forEach((f) => {
+          const item = {};
+          item.title = l.get('title');
+          item.feature = f;
+          item.content = getAttributes(f, l);
+          result.push(item);
+        });
+      } else if (collection.length === 1 && queryable) {
+        const item = {};
+        item.title = l.get('title');
+        item.feature = collection[0];
+        item.content = getAttributes(collection[0], l);
+        item.layer = l;
+        result.push(item);
+      }
+    } else if (queryable) {
+      const item = {};
+      item.title = l.get('title');
+      item.feature = feature;
+      item.content = getAttributes(feature, l);
+      item.layer = l;
+      result.push(item);
+    }
+
+    return false;
+  }, {
+    hitTolerance: featureInfo.getHitTolerance()
+  });
+
+  if (cluster) {
+    return false;
+  }
+  return result;
+}
+
+export default {
+  getFeaturesFromRemote,
+  getFeaturesAtPixel
+};
