@@ -1,8 +1,7 @@
-'use strict';
+import Image from 'ol/layer/image';
 
-var ol = require('openlayers');
-
-module.exports = function image(options, source) {
-  options.source = source;
-  return new ol.layer.Image(options);
-};
+export default function image(options, source) {
+  const opt = options;
+  opt.source = source;
+  return new Image(opt);
+}
