@@ -1,19 +1,28 @@
-"use strict";
+import $ from 'jquery';
+import VectorSource from 'ol/source/vector';
+import TopoJSONFormat from 'ol/format/topojson';
+import viewer from '../viewer';
+import vector from './vector';
+import isUrl from '../utils/isurl';
 
-var ol = require('openlayers');
-var $ = require('jquery');
-var viewer = require('../viewer');
-var vector = require('./vector');
-var isUrl = require('../utils/isurl');
+function createSource(options) {
+  return new VectorSource({
+    attributions: options.attribution,
+    url: options.url,
+    format: new TopoJSONFormat({
+      defaultDataProjection: options.projectionCode
+    })
+  });
+}
 
-var topojson = function topojson(layerOptions) {
-  var baseUrl = viewer.getBaseUrl();
-  var topojsonDefault = {
+const topojson = function topojson(layerOptions) {
+  const baseUrl = viewer.getBaseUrl();
+  const topojsonDefault = {
     layerType: 'vector'
   };
-  var topojsonOptions = $.extend(topojsonDefault, layerOptions);
-  var topojsonSource;
-  var sourceOptions = {};
+  const topojsonOptions = $.extend(topojsonDefault, layerOptions);
+  const sourceOptions = {};
+
   sourceOptions.attribution = topojsonOptions.attribution;
   sourceOptions.projectionCode = viewer.getProjectionCode();
   sourceOptions.sourceName = layerOptions.source;
@@ -24,18 +33,8 @@ var topojson = function topojson(layerOptions) {
     sourceOptions.url = baseUrl + topojsonOptions.source;
   }
 
-  topojsonSource = createSource(sourceOptions);
+  const topojsonSource = createSource(sourceOptions);
   return vector(topojsonOptions, topojsonSource);
+};
 
-  function createSource(options) {
-    return new ol.source.Vector({
-      attributions: options.attribution,
-      url: options.url,
-      format: new ol.format.TopoJSON({
-        defaultDataProjection: options.projectionCode
-      })
-    });
-  }
-}
-
-module.exports = topojson;
+export default topojson;
