@@ -1,20 +1,46 @@
-"use strict";
+import $ from 'jquery';
+import TopoJSONFormat from 'ol/format/topojson';
+import GeoJSONFormat from 'ol/format/geojson';
+import MVTFormat from 'ol/format/mvt';
+import VectorTileSource from 'ol/source/vectortile';
+import viewer from '../viewer';
+import vector from './vector';
+import maputils from '../maputils';
 
-var ol = require('openlayers');
-var $ = require('jquery');
-var viewer = require('../viewer');
-var vector = require('./vector');
-var maputils = require('../maputils');
+function createSource(opt, vectortileOptions) {
+  const options = opt;
+  let format;
+  switch (vectortileOptions.format) {
+    case 'topojson':
+      format = new TopoJSONFormat();
+      break;
+    case 'geojson':
+      format = new GeoJSONFormat();
+      break;
+    case 'pbf':
+      format = new MVTFormat();
+      break;
+    default:
+      break;
+  }
+  if (vectortileOptions.layerURL) {
+    options.url += vectortileOptions.layerURL;
+  } else {
+    options.url += `${vectortileOptions.layerName}@${vectortileOptions.gridset}@${vectortileOptions.format}/{z}/{x}/{-y}.${vectortileOptions.format}`;
+  }
+  options.format = format;
+  return new VectorTileSource(options);
+}
 
-var vectortile = function vectortile(layerOptions) {
-  var vectortileDefault = {
+const vectortile = function vectortile(layerOptions) {
+  const vectortileDefault = {
     layerType: 'vectortile',
     featureinfoLayer: undefined
   };
-  var sourceDefault = {};
-  var vectortileOptions = $.extend(vectortileDefault, layerOptions);
+  const sourceDefault = {};
+  const vectortileOptions = $.extend(vectortileDefault, layerOptions);
   vectortileOptions.sourceName = vectortileOptions.name;
-  var sourceOptions = $.extend(sourceDefault, viewer.getMapSource()[layerOptions.source]);
+  const sourceOptions = $.extend(sourceDefault, viewer.getMapSource()[layerOptions.source]);
   sourceOptions.attributions = vectortileOptions.attribution;
   sourceOptions.projection = viewer.getProjectionCode() || 'EPSG:3857';
 
@@ -30,31 +56,8 @@ var vectortile = function vectortile(layerOptions) {
     }
   }
 
-  var vectortileSource = createSource(sourceOptions, vectortileOptions);
+  const vectortileSource = createSource(sourceOptions, vectortileOptions);
   return vector(vectortileOptions, vectortileSource);
-}
+};
 
-function createSource(options, vectortileOptions) {
-  var format;
-  switch(vectortileOptions.format) {
-    case 'topojson':
-    format = new ol.format.TopoJSON();
-    break;
-    case 'geojson':
-    format = new ol.format.GeoJSON();
-    break;
-    case 'pbf':
-    format = new ol.format.MVT();
-    break;
-  }
-  if(vectortileOptions.layerURL){
-    options.url += vectortileOptions.layerURL;
-  }
-  else {
-    options.url += vectortileOptions.layerName + '@'  + vectortileOptions.gridset + '@' + vectortileOptions.format + '/{z}/{x}/{-y}.'+ vectortileOptions.format;
-  }
-  options.format = format;
-  return new ol.source.VectorTile(options);
-}
-
-module.exports = vectortile;
+export default vectortile;
