@@ -1,59 +1,59 @@
-"use strict";
+import WMTSSource from 'ol/source/wmts';
+import Tilegrid from 'ol/tilegrid/wmts';
+import Extent from 'ol/extent';
+import $ from 'jquery';
+import viewer from '../viewer';
+import tile from './tile';
 
-var ol = require('openlayers');
-var $ = require('jquery');
-var viewer = require('../viewer');
-var tile = require('./tile');
+function createSource(options) {
+  return new WMTSSource({
+    crossOrigin: 'anonymous',
+    attributions: options.attribution,
+    url: options.url,
+    projection: options.projectionCode,
+    layer: options.id,
+    matrixSet: options.matrixSet,
+    format: options.format,
+    tileGrid: new Tilegrid({
+      origin: options.origin || Extent.getTopLeft(options.projectionExtent),
+      resolutions: options.resolutions,
+      matrixIds: options.matrixIds,
+      tileSize: options.tileSize
+    }),
+    style: 'raster'
+  });
+}
 
-var wmts = function wmts(layerOptions) {
-  var wmtsDefault = {
+const wmts = function wmts(layerOptions) {
+  const wmtsDefault = {
     layerType: 'tile',
     featureinfoLayer: undefined
   };
-  var sourceDefault = {
+  const sourceDefault = {
     matrixSet: viewer.getProjectionCode(),
-    matrixIdsPrefix: viewer.getProjectionCode() + ':',
+    matrixIdsPrefix: `${viewer.getProjectionCode()}:`,
     format: 'image/png',
     resolutions: viewer.getResolutions(),
     tileSize: [256, 256]
   };
-  var wmtsOptions = $.extend(wmtsDefault, layerOptions);
+  const wmtsOptions = $.extend(wmtsDefault, layerOptions);
   wmtsOptions.name.split(':').pop();
   wmtsOptions.sourceName = wmtsOptions.name;
-  var sourceOptions = $.extend(sourceDefault, viewer.getMapSource()[layerOptions.source]);
-  if (wmtsOptions.hasOwnProperty('format')) {
+  const sourceOptions = $.extend(sourceDefault, viewer.getMapSource()[layerOptions.source]);
+  if (Object.prototype.hasOwnProperty.call(wmtsOptions, 'format')) {
     sourceOptions.format = wmtsOptions.format;
   }
   sourceOptions.attribution = wmtsOptions.attribution;
   sourceOptions.projectionCode = viewer.getProjectionCode();
   sourceOptions.matrixIds = [];
-  sourceOptions.resolutions.forEach(function(resolution, i) {
+  sourceOptions.resolutions.forEach((resolution, i) => {
     sourceOptions.matrixIds[i] = sourceOptions.matrixIdsPrefix + i;
   });
   sourceOptions.projectionExtent = viewer.getProjection().getExtent();
   sourceOptions.id = wmtsOptions.id;
 
-  var wmtsSource = createSource(sourceOptions);
+  const wmtsSource = createSource(sourceOptions);
   return tile(wmtsOptions, wmtsSource);
+};
 
-  function createSource(options) {
-    return new ol.source.WMTS({
-      crossOrigin: 'anonymous',
-      attributions: options.attribution,
-      url: options.url,
-      projection: options.projectionCode,
-      layer: options.id,
-      matrixSet: options.matrixSet,
-      format: options.format,
-      tileGrid: new ol.tilegrid.WMTS({
-        origin: options.origin || ol.extent.getTopLeft(options.projectionExtent),
-        resolutions: options.resolutions,
-        matrixIds: options.matrixIds,
-        tileSize: options.tileSize
-      }),
-      style: 'raster'
-    })
-  }
-}
-
-module.exports = wmts;
+export default wmts;
