@@ -1,26 +1,11 @@
-"use strict";
+import $ from 'jquery';
+import viewer from './viewer';
+import mapUtils from './maputils';
+import group from './layer/group';
+import type from './layer/layertype';
 
-var $ = require('jquery');
-var viewer = require('./viewer');
-var mapUtils = require('./maputils');
-var group = require('./layer/group');
-var type = {};
-var layerCreator;
-type.WFS = require('./layer/wfs');
-//type.AGS_FEATURE = require('./layer/agsfeature');
-//type.TOPOJSON = require('./layer/topojson');
-//type.GEOJSON = require('./layer/geojson');
-type.WMS = require('./layer/wms');
-type.WMTS = require('./layer/wmts');
-//type.AGS_TILE = require('./layer/agstile');
-//type.XYZ = require('./layer/xyz');
-//type.OSM = require('./layer/osm');
-//type.VECTORTILE = require('./layer/vectortile');
-//type.FEATURE = require('./layer/featurelayer');
-type.GROUP = groupLayer;
-
-layerCreator = function layerCreator(opt_options) {
-  var defaultOptions = {
+const layerCreator = function layerCreator(optOptions) {
+  const defaultOptions = {
     name: undefined,
     id: undefined,
     title: undefined,
@@ -44,12 +29,12 @@ layerCreator = function layerCreator(opt_options) {
     attributes: undefined,
     tileSize: viewer.getTileSize()
   };
-  var projection = viewer.getProjection();
-  var options = opt_options || {};
-  var layerOptions = $.extend(defaultOptions, options);
-  var name = layerOptions.name;
-  layerOptions.minResolution = layerOptions.hasOwnProperty('minScale') ? mapUtils.scaleToResolution(layerOptions.minScale, projection): undefined;
-  layerOptions.maxResolution = layerOptions.hasOwnProperty('maxScale') ? mapUtils.scaleToResolution(layerOptions.maxScale, projection): undefined;
+  const projection = viewer.getProjection();
+  const options = optOptions || {};
+  const layerOptions = $.extend(defaultOptions, options);
+  const name = layerOptions.name;
+  layerOptions.minResolution = Object.prototype.hasOwnProperty.call(layerOptions, 'minScale') ? mapUtils.scaleToResolution(layerOptions.minScale, projection) : undefined;
+  layerOptions.maxResolution = Object.prototype.hasOwnProperty.call(layerOptions, 'maxScale') ? mapUtils.scaleToResolution(layerOptions.maxScale, projection) : undefined;
   layerOptions.extent = layerOptions.extent || viewer.getExtent();
   layerOptions.sourceName = layerOptions.source;
   layerOptions.styleName = layerOptions.style;
@@ -59,28 +44,29 @@ layerCreator = function layerCreator(opt_options) {
 
   layerOptions.name = name.split(':').pop();
 
-  if (type.hasOwnProperty(layerOptions.type)) {
+  if (Object.prototype.hasOwnProperty.call(type, layerOptions.type)) {
     return type[layerOptions.type](layerOptions, layerCreator);
-  } else {
-    console.log('Layer type is missing or layer type is not correct. Check your layer definition: ');
-    console.log(layerOptions);
   }
-}
+
+  console.log('Layer type is missing or layer type is not correct. Check your layer definition: ');
+  console.log(layerOptions);
+
+  return false;
+};
 
 function groupLayer(options) {
-  var layers;
-  var layerOptions;
-  if (options.hasOwnProperty('layers')) {
-    layers = options.layers.map(function(layer) {
-      return layerCreator(layer);
-    });
+  if (Object.prototype.hasOwnProperty.call(options, 'layers')) {
+    const layers = options.layers.map(layer => layerCreator(layer));
 
-    layerOptions = {};
+    const layerOptions = {};
     layerOptions.layers = layers;
     return group($.extend(options, layerOptions));
-  } else {
-    console.log('Group layer has no layers');
   }
+
+  console.log('Group layer has no layers');
+  return false;
 }
 
-module.exports = layerCreator;
+type.GROUP = groupLayer;
+
+export default layerCreator;
