@@ -1,21 +1,38 @@
-"use strict";
-var $ = require('jquery');
-var supports = require('./utils/supports');
-var permalink = require('./permalink/permalink');
-var getUrl = require('./utils/geturl');
-var isUrl = require('./utils/isurl');
-var trimUrl = require('./utils/trimurl');
+import $ from 'jquery';
+import supports from './utils/supports';
+import permalink from './permalink/permalink';
+import getUrl from './utils/geturl';
+import isUrl from './utils/isurl';
+import trimUrl from './utils/trimurl';
 
-var mapLoader = function (mapOptions, config) {
-  var map = {};
-  var mapEl = config.target;
-  var format = 'json';
-  var cors = config.crossOrigin;
+function loadSvgSprites(baseUrl, config) {
+  const svgSprites = config.svgSprites;
+  const svgPath = config.svgSpritePath;
+  const svgPromises = [];
+  svgSprites.forEach((sprite) => {
+    const promise = $.get(baseUrl + svgPath + sprite, (data) => {
+      const div = document.createElement('div');
+      div.innerHTML = new XMLSerializer().serializeToString(data.documentElement);
+      document.body.insertBefore(div, document.body.childNodes[0]);
+    });
+    svgPromises.push(promise);
+    return svgPromises;
+  });
+}
 
-  var urlParams = undefined;
+const mapLoader = function mapLoader(mapOptions, config) {
+  const map = {};
+  let mapEl = config.target;
+  const format = 'json';
+  let urlParams;
+  let url;
+  let mapUrl;
+  let baseUrl;
+  let json;
+
 
   if (mapEl.substring(0, 1) !== '#') {
-    mapEl = '#' + mapEl;
+    mapEl = `#${mapEl}`;
   }
   map.el = mapEl;
 
@@ -24,24 +41,12 @@ var mapLoader = function (mapOptions, config) {
     return undefined;
   }
 
-  // Check if authorization is required before map options is loaded
-  if (config.authorizationUrl) {
-    return $.ajax({
-        url: config.authorizationUrl
-      })
-      .then(function(data) {
-        return loadMapOptions();
-      });
-  } else {
-    return loadMapOptions();
-  }
-
   function loadMapOptions() {
-    if (typeof(mapOptions) === 'object') {
+    if (typeof (mapOptions) === 'object') {
       if (window.location.hash) {
         urlParams = permalink.parsePermalink(window.location.href);
       }
-      var baseUrl = config.baseUrl || '';
+      baseUrl = config.baseUrl || '';
       map.options = $.extend(config, mapOptions);
       if (mapOptions.controls) {
         map.options.controls = config.defaultControls.concat(mapOptions.controls);
@@ -54,74 +59,65 @@ var mapLoader = function (mapOptions, config) {
       map.options.baseUrl = baseUrl;
       map.options.config = mapOptions;
 
-      return $.when.apply($, loadSvgSprites(baseUrl, config))
-        .then(function(sprites) {
-          return map;
-        });
-
-    } else if (typeof(mapOptions) === 'string') {
+      return $.when(loadSvgSprites(baseUrl, config))
+        .then(() => map);
+    } else if (typeof (mapOptions) === 'string') {
       if (isUrl(mapOptions)) {
         urlParams = permalink.parsePermalink(mapOptions);
-        var url = mapOptions.split('#')[0];
-        var mapUrl = url;
+        url = mapOptions.split('#')[0];
+        mapUrl = url;
 
-        //remov file name if included in
+        // remove file name if included in
         url = trimUrl(url);
 
-        var baseUrl = config.baseUrl || url;
-        var json = urlParams.map + '.json';
+        baseUrl = config.baseUrl || url;
+
+        json = `${urlParams.map}.json`;
         url += json;
       } else {
-        var json = mapOptions;
+        json = mapOptions;
         if (window.location.hash) {
           urlParams = permalink.parsePermalink(window.location.href);
           if (urlParams.map) {
-            json = urlParams.map + '.json';
+            json = `${urlParams.map}.json`;
           }
         }
-        var baseUrl = config.baseUrl || '';
-        var url = baseUrl + json;
-        var mapUrl = getUrl();
+        baseUrl = config.baseUrl || '';
+        url = baseUrl + json;
+        mapUrl = getUrl();
       }
 
-      return $.when.apply($, loadSvgSprites(baseUrl, config))
-        .then(function(sprites) {
-          return $.ajax({
-              url: url,
-              dataType: format
-            })
-            .then(function(data) {
-              map.options = $.extend(config, data);
-              if (data.controls) {
-                map.options.controls = config.defaultControls.concat(data.controls);
-              } else {
-                map.options.controls = config.defaultControls;
-              }
-              map.options.url = mapUrl;
-              map.options.map = json;
-              map.options.params = urlParams;
-              map.options.baseUrl = baseUrl;
-              map.options.config = data;
-              return map;
-            });
-        });
+      return $.when(loadSvgSprites(baseUrl, config))
+        .then(() => $.ajax({
+          url,
+          dataType: format
+        })
+          .then((data) => {
+            map.options = $.extend(config, data);
+            if (data.controls) {
+              map.options.controls = config.defaultControls.concat(data.controls);
+            } else {
+              map.options.controls = config.defaultControls;
+            }
+            map.options.url = mapUrl;
+            map.options.map = json;
+            map.options.params = urlParams;
+            map.options.baseUrl = baseUrl;
+            map.options.config = data;
+            return map;
+          }));
     }
+    return null;
   }
-}
 
-function loadSvgSprites(baseUrl, config) {
-  var svgSprites = config.svgSprites;
-  var svgPath = config.svgSpritePath;
-  var svgPromises = [];
-  svgSprites.forEach(function(sprite) {
-    var promise = $.get(baseUrl + svgPath + sprite, function(data) {
-        var div = document.createElement("div");
-        div.innerHTML = new XMLSerializer().serializeToString(data.documentElement);
-        document.body.insertBefore(div, document.body.childNodes[0]);
-      });
-    svgPromises.push(promise);
-    return svgPromises;
-  });
-}
+  // Check if authorization is required before map options is loaded
+  if (config.authorizationUrl) {
+    return $.ajax({
+      url: config.authorizationUrl
+    })
+      .then(() => loadMapOptions());
+  }
+  return loadMapOptions();
+};
 
-module.exports = mapLoader;
+export default mapLoader;
