@@ -1,14 +1,10 @@
-"use strict";
+const pin = [{
+  icon: {
+    anchor: [0.5, 32],
+    anchorXUnits: 'fraction',
+    anchorYUnits: 'pixels',
+    src: 'img/png/drop_blue.png'
+  }
+}];
 
-var pin = [
-    {
-      'icon': {
-          anchor: [0.5, 32],
-          anchorXUnits: 'fraction',
-          anchorYUnits: 'pixels',
-          src: 'img/png/drop_blue.png'
-      }
-    }
-];
-
-module.exports = pin;
+export default pin;
