@@ -1,115 +1,97 @@
-var ol = require('openlayers');
-var getColor = require('../getcolor');
+import Circle from 'ol/style/circle';
+import Fill from 'ol/style/fill';
+import Stroke from 'ol/style/stroke';
+import Style from 'ol/style/style';
+import getColor from '../getcolor';
 
-module.exports = function defaultStyle(params) {
-  var fill = new ol.style.Fill({
+export default function defaultStyle() {
+  const fill = new Fill({
     color: ''
   });
-  var stroke = new ol.style.Stroke({
+  const stroke = new Stroke({
     color: '',
     width: 0.1,
     lineCap: 'square',
     lineJoin: 'round'
   });
-  var noStroke = new ol.style.Stroke({
-    color: '',
-    width: 0.0
-  });
-  var overlayedStroke = new ol.style.Stroke({
-    color: '',
-    width: 0.1,
-    lineCap: 'square',
-    lineJoin: 'round'
-  });
-  var dashedStroke = new ol.style.Stroke({
-    color: '',
-    width: 1,
-    lineDash: [1, 2]
-  });
-  var polygon = new ol.style.Style({
-    fill: fill,
+  const polygon = new Style({
+    fill,
     zIndex: 1
   });
-  var strokedPolygon = new ol.style.Style({
-    fill: fill,
-    stroke: stroke,
+  const strokedPolygon = new Style({
+    fill,
+    stroke,
     zIndex: 2
   });
-  var dashedPolygon = new ol.style.Style({
-    fill: fill,
-    stroke: dashedStroke,
-    zIndex: 2
-  });
-  var line = new ol.style.Style({
-    stroke: stroke,
+  const line = new Style({
+    stroke,
     zIndex: 10
   });
-  var overlayedLine = new ol.style.Style({
-    stroke: overlayedStroke,
-    zIndex: 11
-  });
-  var dashedLine = new ol.style.Style({
-    stroke: dashedStroke,
-    zIndex: 12
-  });
-  var point = new ol.style.Style({
-    image: new ol.style.Circle({
+  const point = new Style({
+    image: new Circle({
       radius: 5,
-      fill: fill,
-      stroke: stroke
+      fill,
+      stroke
     }),
     zIndex: 50
   });
-  var styles = [];
+  const styles = [];
 
-  return function(feature, resolution) {
+  return function style(feature) {
     polygon.setZIndex(1);
     line.setZIndex(10);
-    var length = 0;
-    var geom = feature.getGeometry().getType();
+    let length = 0;
+    const geom = feature.getGeometry().getType();
     switch (geom) {
       case 'Polygon':
         stroke.setColor(getColor('blue'));
         stroke.setWidth(1);
         fill.setColor(getColor('blue', 0.8));
-        styles[length++] = strokedPolygon;
+        styles[length] = strokedPolygon;
+        length += 1;
         break;
       case 'MultiPolygon':
         stroke.setColor(getColor('blue'));
         stroke.setWidth(1);
         fill.setColor(getColor('blue', 0.8));
-        styles[length++] = strokedPolygon;
+        styles[length] = strokedPolygon;
+        length += 1;
         break;
       case 'LineString':
         stroke.setColor(getColor('red'));
         stroke.setWidth(1);
-        styles[length++] = line;
+        styles[length] = line;
+        length += 1;
         break;
       case 'MultiLineString':
         stroke.setColor(getColor('red'));
         stroke.setWidth(1);
-        styles[length++] = line;
+        styles[length] = line;
+        length += 1;
         break;
       case 'Point':
         stroke.setColor(getColor('yellow'));
         stroke.setWidth(1);
         fill.setColor(getColor('yellow', 0.8));
-        styles[length++] = point;
+        styles[length] = point;
+        length += 1;
         break;
       case 'MultiPoint':
         stroke.setColor(getColor('yellow'));
         stroke.setWidth(1);
         fill.setColor(getColor('yellow', 0.8));
-        styles[length++] = point;
+        styles[length] = point;
+        length += 1;
         break;
       default:
         stroke.setColor(getColor('blue'));
         stroke.setWidth(1);
         fill.setColor(getColor('blue', 0.8));
-        styles[length++] = strokedPolygon;
+        styles[length] = strokedPolygon;
+        length += 1;
         break;
     }
     styles.length = length;
     return styles;
   };
-};
+}
