@@ -1,21 +1,18 @@
-"use strict";
-
-
 /**
 Function to generate a unique id.
 From Briguy37 http://stackoverflow.com/questions/105034/create-guid-uuid-in-javascript
 */
-var generateUUID = function generateUUID() {
-  var d = new Date().getTime();
-  if (window.performance && typeof window.performance.now === "function") {
-    d += performance.now(); //use high-precision timer if available
+const generateUUID = function generateUUID() {
+  let d = new Date().getTime();
+  if (window.performance && typeof window.performance.now === 'function') {
+    d += performance.now(); // use high-precision timer if available
   }
-  var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-    var r = (d + Math.random() * 16) % 16 | 0;
+  const uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+    const r = (d + Math.random() * 16) % 16 | 0;
     d = Math.floor(d / 16);
-    return (c == 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+    return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
   });
   return uuid;
-}
+};
 
-module.exports = generateUUID;
+export default generateUUID;
