@@ -1,8 +1,8 @@
-"use strict";
-
-module.exports = function getUrl() {
-    return location.protocol +'//'+
-            location.hostname +
-            (location.port?":"+location.port:"") +
-            location.pathname;
+export default function getUrl() {
+  return [
+    `${window.location.protocol}//`,
+    `${window.location.hostname}`,
+    (window.location.port ? `:${window.location.port}` : ''),
+    `${window.location.pathname}`
+  ].join('');
 }
