@@ -1,21 +1,19 @@
-"use strict";
-
-module.exports = function numberFormatter(numberToFormat) {
-  var nr = numberToFormat;
-  var length = nr.toString().length;
+export default function numberFormatter(numberToFormat) {
+  let nr = numberToFormat;
+  const length = nr.toString().length;
   if (length > 3) {
-    var delimiter = ' '; //white space as delimiter
+    const delimiter = ' '; // white space as delimiter
 
-    //Round by factor
+    // Round by factor
     if (nr >= 1000) {
-      var factor = Math.pow(10, length) / 10000;
+      const factor = (10 ** length) / 10000;
       nr = Math.round(nr / factor) * factor;
     }
 
-    var nrStr = nr.toString();
-    var a = [];
+    let nrStr = nr.toString();
+    const a = [];
     while (nrStr.length > 3) {
-      var n = nrStr.substr(nrStr.length - 3);
+      const n = nrStr.substr(nrStr.length - 3);
       a.unshift(n);
       nrStr = nrStr.substr(0, nrStr.length - 3);
     }
@@ -24,7 +22,6 @@ module.exports = function numberFormatter(numberToFormat) {
     }
     nrStr = a.join(delimiter);
     return nrStr;
-  } else {
-    return nr.toString();
   }
+  return nr.toString();
 }
