@@ -1,61 +1,54 @@
-"use strict";
+const replacer = function replacer() {
+  let start;
+  let end;
+  let helper;
+  let helperNS;
+  let helperArg;
 
-var Replacer = function() {
-
-  var start = undefined;
-  var end = undefined;
-  var helper = undefined;
-  var helperNS = undefined;
-  var helperArg = undefined;
-
-  function replace(name, obj, options) {
-      start = options.start || '{{';
-      end = options.end || '}}';
-      helper = options.helper || {};
-      helperNS = options.helperNS || '@';
-      helperArg = [options.helperArg] || [];
-
-      var result = searchAndReplace(name, obj);
-      return result;
+  function getArgs(str) {
+    const args = str.match(/\((.*?)\)/);
+    if (args) {
+      return [args[1].split(','), str.substring(0, args.index)];
+    }
+    return ['', str];
   }
 
   function searchAndReplace(name, obj) {
-      var regex = new RegExp(start + '(.*?)' +  end, "g");
-      var matches = regex.exec(name);
-      if(matches) {
-          var val = obj.hasOwnProperty(matches[1]) ? obj[matches[1]] : '';
-          if (!val) {
-            var nsIndex = matches[0].indexOf(helperNS);
-            if (nsIndex) {
-              var helperParts = getArgs(matches[1]);
-              var helperName = helperParts[1].substring(nsIndex - 1);
-              var args = helperArg.concat(helperParts[0]);
-              val = helper.hasOwnProperty(helperName) ? helper[helperName].apply(null, args).toString() : '';
-            }
-          }
-          return searchAndReplace(name.replace(matches[0], val), obj);
+    const regex = new RegExp(`${start}(.*?)${end}`, 'g');
+    const matches = regex.exec(name);
+    if (matches) {
+      let val = Object.prototype.hasOwnProperty.call(obj, matches[1]) ? obj[matches[1]] : '';
+      if (!val) {
+        const nsIndex = matches[0].indexOf(helperNS);
+        if (nsIndex) {
+          const helperParts = getArgs(matches[1]);
+          const helperName = helperParts[1].substring(nsIndex - 1);
+          const args = helperArg.concat(helperParts[0]);
+          val = Object.prototype.hasOwnProperty.call(helper, helperName) ? helper[helperName].apply(null, args).toString() : '';
+        }
       }
-      else {
-          return name;
-      }
-  }
-
-  function getArgs(str) {
-    var args = str.match(/\((.*?)\)/);
-    if (args) {
-      return [args[1].split(','), str.substring(0, args.index)];
-    } else {
-      return ['',str];
+      return searchAndReplace(name.replace(matches[0], val), obj);
     }
+    return name;
   }
 
-	return {
-  		replace: function(name, obj, options) {
-          var opt = options || {};
-          return (replace(name, obj, opt));
-  		}
-	};
+  function replace(name, obj, options) {
+    start = options.start || '{{';
+    end = options.end || '}}';
+    helper = options.helper || {};
+    helperNS = options.helperNS || '@';
+    helperArg = [options.helperArg] || [];
 
-}
+    const result = searchAndReplace(name, obj);
+    return result;
+  }
 
-module.exports = Replacer();
+  return {
+    replace(name, obj, options) {
+      const opt = options || {};
+      return (replace(name, obj, opt));
+    }
+  };
+};
+
+export default replacer();
