@@ -1,13 +1,11 @@
-"use strict";
-var renderError = require('./rendererror');
+import renderError from './rendererror';
 
-module.exports = function supports(type, el) {
-  var canvas = document.createElement('canvas').getContext;
-  var requestAnimationFrame = window.requestAnimationFrame;
+export default function supports(type, el) {
+  const canvas = document.createElement('canvas').getContext;
+  const requestAnimationFrame = window.requestAnimationFrame;
   if (!!canvas && !!requestAnimationFrame) {
     return true;
-  } else {
-    renderError('browser', el);
-    return false;
   }
+  renderError('browser', el);
+  return false;
 }
