@@ -1,8 +1,7 @@
-"use strict";
-var isUrl = require('./isurl');
+import isUrl from './isurl';
 
-//Checks if str is a valid url. If not append str to baseUrl
-module.exports = function validateUrl(str, baseUrl) {
-   var url = isUrl(str) ? str : baseUrl + str;
-   return url;
+// Checks if str is a valid url. If not append str to baseUrl
+export default function validateUrl(str, baseUrl) {
+  const url = isUrl(str) ? str : baseUrl + str;
+  return url;
 }
