@@ -1,21 +1,24 @@
-"use strict";
+import proj4 from 'proj4';
+import Map from 'ol/map';
+import View from 'ol/view';
+import Collection from 'ol/collection';
+import Projection from 'ol/proj/projection';
+import project from 'ol/proj';
+import Feature from 'ol/feature';
+import geom from 'ol/geom/geometry';
+import $ from 'jquery';
+import template from './templates/viewertemplate';
+import elQuery from './utils/elquery';
+import featureinfo from './featureinfo';
+import getcenter from './geometry/getcenter';
+import maputils from './maputils';
+import getattributes from './getattributes';
+import style from './style';
+import layerCreator from './layercreator';
+import getCapabilities from './getCapabilities';
 
-var ol = require('openlayers');
-var $ = require('jquery');
-var template = require("./templates/viewer.handlebars");
-var Modal = require('./modal');
-var utils = require('./utils');
-var isUrl = require('./utils/isurl');
-var elQuery = require('./utils/elquery');
-var featureinfo = require('./featureinfo');
-var maputils = require('./maputils');
-var style = require('./style')();
-var layerCreator = require('./layercreator');
-var getCapabilities = require('./getCapabilities');
-
-var map;
-var template;
-var settings = {
+let map;
+const settings = {
   projection: '',
   projectionCode: '',
   projectionExtent: '',
@@ -32,122 +35,64 @@ var settings = {
   featureInfoOverlay: undefined,
   editLayer: null
 };
-var urlParams;
-var pageSettings;
-var pageTemplate = {};
+let urlParams;
+let pageSettings;
+const pageTemplate = {};
 
-function init(el, mapOptions) {
-  render(el, mapOptions);
+function render(el, mapOptions) {
+  pageSettings = mapOptions.pageSettings;
+  pageTemplate.mapClass = 'o-map';
 
-  // Read and set projection
-  if (mapOptions.hasOwnProperty('proj4Defs') && window.proj4) {
-    var proj = mapOptions['proj4Defs'];
-
-    //Register proj4 projection definitions
-    for (var i = 0; i < proj.length; i++) {
-      proj4.defs(proj[i].code, proj[i].projection);
-      if (proj[i].hasOwnProperty('alias')) {
-        proj4.defs(proj[i].alias, proj4.defs(proj[i].code));
+  if (pageSettings) {
+    if (pageSettings.footer) {
+      if ('img' in pageSettings.footer) {
+        pageTemplate.img = pageSettings.footer.img;
+      }
+      if ('text' in pageSettings.footer) {
+        pageTemplate.text = pageSettings.footer.text;
+      }
+      if ('url' in pageSettings.footer) {
+        pageTemplate.url = pageSettings.footer.url;
+      }
+      if ('urlText' in pageSettings.footer) {
+        pageTemplate.urlText = pageSettings.footer.urlText;
+      }
+    }
+    if (pageSettings.mapGrid) {
+      if ('visible' in pageSettings.mapGrid && pageSettings.mapGrid.visible === true) {
+        pageTemplate.mapClass = 'o-map o-map-grid';
       }
     }
   }
-  settings.params = urlParams = mapOptions.params || {};
-  settings.map = mapOptions.map;
-  settings.url = mapOptions.url;
-  settings.target = mapOptions.target;
-  settings.baseUrl = mapOptions.baseUrl;
-  settings.breakPoints = mapOptions.breakPoints;
-  settings.extent = mapOptions.extent || undefined;
-  settings.center = urlParams.center || mapOptions.center;
-  settings.zoom = urlParams.zoom || mapOptions.zoom;
-  mapOptions.tileGrid = mapOptions.tileGrid || {};
-  settings.tileSize = mapOptions.tileGrid.tileSize ? [mapOptions.tileGrid.tileSize,mapOptions.tileGrid.tileSize] : [256,256];
-  settings.alignBottomLeft = mapOptions.tileGrid.alignBottomLeft;
-  settings.config = mapOptions.config;
 
-  if (mapOptions.hasOwnProperty('proj4Defs') || mapOptions.projectionCode=="EPSG:3857" || mapOptions.projectionCode=="EPSG:4326") {
-    // Projection to be used in map
-    settings.projectionCode = mapOptions.projectionCode || undefined;
-    settings.projectionExtent = mapOptions.projectionExtent;
-    settings.projection = new ol.proj.Projection({
-      code: settings.projectionCode,
-      extent: settings.projectionExtent,
-      units: getUnits(settings.projectionCode)
-    });
-    settings.resolutions = mapOptions.resolutions || undefined;
-    settings.tileGrid = maputils.tileGrid(settings);
-  }
-
-  settings.source = mapOptions.source;
-  settings.groups = mapOptions.groups;
-  settings.editLayer = mapOptions.editLayer;
-  settings.styles = mapOptions.styles;
-  settings.capabilitiesURL = mapOptions.capabilitiesURL;
-  settings.clusterOptions = mapOptions.clusterOptions || {};
-  style.init();
-  settings.controls = mapOptions.controls;
-  settings.consoleId = mapOptions.consoleId || 'o-console';
-  settings.featureinfoOptions = mapOptions.featureinfoOptions || {};
-  settings.enableRotation = mapOptions.enableRotation === false ? false : true;
-
-  //If url arguments, parse this settings
-  if (window.location.search) {
-    parseArg();
-  }
-
-  loadMap();
-  settings.layers = createLayers(mapOptions.layers, urlParams.layers);
-  addLayers(settings.layers);
-
-  elQuery(map, {
-    breakPoints: mapOptions.breakPoints,
-    breakPointsPrefix: mapOptions.breakPointsPrefix,
-  });
-
-  if (urlParams.pin) {
-    settings.featureinfoOptions.savedPin = urlParams.pin;
-  }
-
-  //This needs further development for proper handling in permalink
-  else if (urlParams.selection) {
-    settings.featureinfoOptions.savedSelection = new ol.Feature({
-      geometry: new ol.geom[urlParams.selection.geometryType](urlParams.selection.coordinates)
-    });
-  }
-  featureinfo.init(settings.featureinfoOptions);
-}
-
-function addLayers(layers) {
-  layers.forEach(function(layer) {
-    map.addLayer(layer);
-  });
+  $(el).html(template(pageTemplate));
 }
 
 function createLayers(layerlist, savedLayers) {
-  var layers = [];
-  var getCapabilitiesLayers = [];
+  const layers = [];
+  let getCapabilitiesLayers = [];
 
-  if (settings.capabilitiesURL != undefined) {
-      getCapabilitiesLayers = getCapabilities(settings.capabilitiesURL);
-
-      layerlist.forEach(function(layer) {
-        if (getCapabilitiesLayers.indexOf(layer.name) >= 0) {
+  if (settings.capabilitiesURL !== undefined) {
+    getCapabilitiesLayers = getCapabilities(settings.capabilitiesURL);
+    layerlist.forEach((layer) => {
+      if (getCapabilitiesLayers.indexOf(layer.name) >= 0) {
         layer.secure = false;
       } else {
-        if (layer['secure'] === undefined) {
+        if (layer.secure === undefined) {
           layer.secure = true;
-          layer.abstract = "Du saknar behörighet att se detta lager";
-        };
+          layer.abstract = 'Du saknar behörighet att se detta lager';
+        }
       }
-      })
+    });
   } else {
-    layerlist.forEach(function(layer) {
-      if (layer['secure'] === undefined) {layer.secure = false};
-    })
+    layerlist.forEach((layer) => {
+      if (layer.secure === undefined) { layer.secure = false; }
+    });
   }
 
-  for (var i = layerlist.length - 1; i >= 0; i--) {
-    var savedLayer = {};
+  let i;
+  for (i = layerlist.length - 1; i >= 0; i -= 1) {
+    let savedLayer = {};
     if (savedLayers) {
       savedLayer = savedLayers[layerlist[i].name.split(':').pop()] || {
         visible: false,
@@ -155,17 +100,32 @@ function createLayers(layerlist, savedLayers) {
       };
       savedLayer.name = layerlist[i].name;
     }
-    var layer = $.extend(layerlist[i], savedLayer);
+    const layer = $.extend(layerlist[i], savedLayer);
     layers.push(layerCreator(layer));
   }
   return layers;
 }
 
+function getUnits(proj) {
+  let units;
+  switch (proj) {
+    case 'EPSG:3857':
+      units = 'm';
+      break;
+    case 'EPSG:4326':
+      units = 'degrees';
+      break;
+    default:
+      units = proj4.defs(proj) ? proj4.defs(proj).units : undefined;
+  }
+  return units;
+}
+
 function loadMap() {
-  map = new ol.Map({
+  map = new Map({
     target: 'o-map',
     controls: [],
-    view: new ol.View({
+    view: new View({
       extent: settings.extent || undefined,
       projection: settings.projection || undefined,
       center: settings.center,
@@ -176,48 +136,10 @@ function loadMap() {
   });
 }
 
-function parseArg() {
-  var str = window.location.search.substring(1);
-  var elements = str.split("&");
-
-  for (var i = 0; i < elements.length; i++) {
-
-    //center coordinates
-    if (i == 0) {
-      var z = elements[i].split(",");
-      settings.center[0] = parseInt(z[0]);
-      settings.center[1] = parseInt(z[1]);
-    } else if (i == 1) {
-      settings.zoom = parseInt(elements[i]);
-    } else if (i == 2) {
-      var l = elements[i].split(";");
-      var layers = settings.layers;
-      var la, match;
-      for (var j = 0; j < layers.length; j++) {
-        match = 0;
-        $.each(l, function(index, el) {
-          la = el.split(",");
-          if (layers[j].get('group')) {
-            if ((layers[j].get('group') == 'background') && (la[0] == layers[j].get('name'))) {
-              layers[j].setVisible(true);
-              match = 1;
-            } else if ((layers[j].get('group') == 'background') && (match == 0)) {
-              layers[j].setVisible(false);
-            } else if (la[0] == layers[j].get('name')) {
-              if (la[1] == 1) {
-                layers[j].set('legend', true);
-                layers[j].setVisible(false);
-              } else {
-                layers[j].set('legend', true);
-                layers[j].setVisible(true);
-              }
-            }
-          }
-        })
-      }
-    }
-  }
-
+function addLayers(layers) {
+  layers.forEach((layer) => {
+    map.addLayer(layer);
+  });
 }
 
 function getSettings() {
@@ -233,7 +155,7 @@ function getBaseUrl() {
 }
 
 function getBreakPoints(size) {
-  return size && settings.breakPoints.hasOwnProperty(size) ? settings.breakPoints[size] : settings.breakPoints;
+  return size && size in settings.breakPoints ? settings.breakPoints[size] : settings.breakPoints;
 }
 
 function getMapName() {
@@ -265,32 +187,32 @@ function getResolutions() {
 }
 
 function getMapUrl() {
-  var layerNames = '';
-  var url;
+  let layerNames = '';
+  let url;
 
-  //delete search arguments if present
+  // delete search arguments if present
   if (window.location.search) {
     url = window.location.href.replace(window.location.search, '?');
   } else {
-    url = window.location.href + '?';
+    url = `${window.location.href}?`;
   }
-  var mapView = map.getView();
-  var center = mapView.getCenter();
-  for (var i = 0; i < 2; i++) {
-    center[i] = parseInt(center[i]); //coordinates in integers
+  const mapView = map.getView();
+  const center = mapView.getCenter();
+  for (let i = 0; i < 2; i += 1) {
+    center[i] = parseInt(center[i], 10); // coordinates in integers
   }
-  var zoom = mapView.getZoom();
-  var layers = map.getLayers();
+  const zoom = mapView.getZoom();
+  const layers = map.getLayers();
 
-  //add layer if visible
-  layers.forEach(function(el) {
-    if (el.getVisible() == true) {
-      layerNames += el.get('name') + ';';
-    } else if (el.get('legend') == true) {
-      layerNames += el.get('name') + ',1;';
+  // add layer if visible
+  layers.forEach((el) => {
+    if (el.getVisible() === true) {
+      layerNames += `${el.get('name')};`;
+    } else if (el.get('legend') === true) {
+      layerNames += `${el.get('name')},1;`;
     }
-  })
-  return url + center + '&' + zoom + '&' + layerNames.slice(0, layerNames.lastIndexOf(";"));
+  });
+  return `${url}${center}&${zoom}&${layerNames.slice(0, layerNames.lastIndexOf(';'))}`;
 }
 
 function getMap() {
@@ -302,51 +224,34 @@ function getLayers() {
 }
 
 function getLayersByProperty(key, val, byName) {
-  var layers = map.getLayers().getArray().filter(function(layer) {
-    if (layer.get(key)) {
-      if (layer.get(key) === val) {
-        return layer;
-      }
-    }
-  });
+  const layers = map.getLayers().getArray().filter(layer => layer.get(key) && layer.get(key) === val);
 
   if (byName) {
-    return layers.map(function(layer) {
-      return layer.get('name');
-    });
-  } else {
-    return layers;
+    return layers.map(layer => layer.get('name'));
   }
+  return layers;
 }
 
 function getLayer(layername) {
-  var layer = $.grep(settings.layers, function(obj) {
-    return (obj.get('name') == layername);
-  });
+  const layer = $.grep(settings.layers, obj => obj.get('name') === layername);
   return layer[0];
 }
 
 function getLayerGroups() {
-  var groups = $.grep(settings.layers, function(obj) {
-    return (obj.get('type') === 'GROUP');
-  });
+  const groups = $.grep(settings.layers, obj => (obj.get('type') === 'GROUP'));
   return groups;
 }
 
 function getQueryableLayers() {
-  var queryableLayers = settings.layers.filter(function(layer) {
-    if (layer.get('queryable') && layer.getVisible()) {
-      return layer;
-    }
-  });
+  const queryableLayers = settings.layers.filter(layer => layer.get('queryable') && layer.getVisible());
   return queryableLayers;
 }
 
 function getSearchableLayers(searchableDefault) {
-  var searchableLayers = [];
-  map.getLayers().forEach(function(layer) {
-    var searchable = layer.get('searchable');
-    var visible = layer.getVisible();
+  const searchableLayers = [];
+  map.getLayers().forEach((layer) => {
+    let searchable = layer.get('searchable');
+    const visible = layer.getVisible();
     searchable = searchable === undefined ? searchableDefault : searchable;
     if (searchable === 'always' || (searchable && visible)) {
       searchableLayers.push(layer.get('name'));
@@ -356,24 +261,11 @@ function getSearchableLayers(searchableDefault) {
 }
 
 function getGroup(group) {
-  var group = $.grep(settings.layers, function(obj) {
-    return (obj.get('group') == group);
-  });
-  return group;
-}
-
-function getGroups(opt) {
-  if(opt == 'top') {
-    return settings.groups;
-  } else if (opt == 'sub') {
-    return getSubgroups();
-  } else {
-    return settings.groups.concat(getSubgroups());
-  }
+  return settings.layers.filter(obj => obj.get('group') === group);
 }
 
 function getSubgroups() {
-  var subgroups = [];
+  const subgroups = [];
 
   function findSubgroups(groups, n) {
     if (n >= groups.length) {
@@ -381,18 +273,27 @@ function getSubgroups() {
     }
 
     if (groups[n].groups) {
-      groups[n].groups.forEach(function(subgroup) {
+      groups[n].groups.forEach((subgroup) => {
         subgroups.push(subgroup);
       });
 
       findSubgroups(groups[n].groups, 0);
     }
 
-    findSubgroups(groups, n+1);
+    findSubgroups(groups, n + 1);
   }
 
   findSubgroups(settings.groups, 0);
   return subgroups;
+}
+
+function getGroups(opt) {
+  if (opt === 'top') {
+    return settings.groups;
+  } else if (opt === 'sub') {
+    return getSubgroups();
+  }
+  return settings.groups.concat(getSubgroups());
 }
 
 function getProjectionCode() {
@@ -408,49 +309,40 @@ function getMapSource() {
 }
 
 function getControlNames() {
-  var controlNames = settings.controls.map(function(obj) {
-    return obj.name;
-  });
-  return controlNames;
+  return settings.controls.map(obj => obj.name);
 }
 
 function getTarget() {
   return settings.target;
 }
 
-function getClusterOptions(){
+function getClusterOptions() {
   return settings.clusterOptions;
 }
 
 function checkScale(scale, maxScale, minScale) {
   if (maxScale || minScale) {
-
     // Alter 1: maxscale and minscale
     if (maxScale && minScale) {
       if ((scale > maxScale) && (scale < minScale)) {
         return true;
       }
-    }
-
-    // Alter 2: only maxscale
-    else if (maxScale) {
+    } else if (maxScale) {
+      // Alter 2: only maxscale
       if (scale > maxScale) {
         return true;
       }
-    }
-
-    // Alter 3: only minscale
-    else if (minScale) {
+    } else if (minScale) {
+      // Alter 3: only minscale
       if (scale < minScale) {
         return true;
       }
     }
-  }
-
-  // Alter 4: no scale limit
-  else {
+  } else {
+    // Alter 4: no scale limit
     return true;
   }
+  return false;
 }
 
 function getConsoleId() {
@@ -458,161 +350,190 @@ function getConsoleId() {
 }
 
 function getScale(resolution) {
-  var dpi = 25.4 / 0.28;
-  var mpu = settings.projection.getMetersPerUnit();
-  var scale = resolution * mpu * 39.37 * dpi;
+  const dpi = 25.4 / 0.28;
+  const mpu = settings.projection.getMetersPerUnit();
+  let scale = resolution * mpu * 39.37 * dpi;
   scale = Math.round(scale);
   return scale;
 }
 
-function getUnits(proj) {
-  var units;
-  switch(proj) {
-    case 'EPSG:3857':
-      units = 'm';
-      break;
-    case 'EPSG:4326':
-      units = 'degrees';
-      break;
-    default:
-      units = proj4.defs(proj) ? proj4.defs(proj).units : undefined;
-  }
-  return units;
-}
-
-function autoPan() {
-  /*Workaround to remove when autopan implemented for overlays */
-  var el = $('.o-popup');
-  var center = map.getView().getCenter();
-  var popupOffset = $(el).offset();
-  var mapOffset = $('#' + map.getTarget()).offset();
-  var offsetY = popupOffset.top - mapOffset.top;
-  var mapSize = map.getSize();
-  var offsetX = (mapOffset.left + mapSize[0]) - (popupOffset.left + $(el).outerWidth(true));
-
-  // Check if mapmenu widget is used and opened
-  var menuSize = 0;
-  if (settings.controls.hasOwnProperty('mapmenu')) {
-    menuSize = settings.controls.mapmenu.getTarget().offset().left > 0 ? mapSize[0] - settings.controls.mapmenu.getTarget().offset().left : menuSize = 0;
-  }
-  if (offsetY < 0 || offsetX < 0 + menuSize || offsetX > (mapSize[0] - $(el).outerWidth(true))) {
-    var dx = 0,
-      dy = 0;
-    if (offsetX < 0 + menuSize) {
-      dx = (-offsetX + menuSize) * map.getView().getResolution();
-    }
-    if (offsetX > (mapSize[0] - $(el).outerWidth(true))) {
-      dx = -($(el).outerWidth(true) - (mapSize[0] - offsetX)) * map.getView().getResolution();
-    }
-    if (offsetY < 0) {
-      dy = (-offsetY) * map.getView().getResolution();
-    }
-    map.getView().animate({
-      center: ([center[0] + dx, center[1] + dy]),
-	  duration:300
-    });
-  }
-  /*End workaround*/
-}
 function removeLayer(name) {
-  var $ul;
-
-  settings.layers.forEach(function(layer, i, obj) {
+  settings.layers.forEach((layer, i, obj) => {
     if (layer.get('name') === name) {
-      obj.splice(i, 1)
+      obj.splice(i, 1);
     }
   });
 
-  $ul = $("#o-mapmenu").find("#" + name).closest('ul');
-  $ul.find('#' + name).remove();
+  const $ul = $('#o-mapmenu').find(`#${name}`).closest('ul');
+  $ul.find(`#${name}`).remove();
   if ($ul.children().length === 1) {
     $ul.remove();
   }
 
-  $('#o-legend-' + name).remove();
-  map.removeLayer(getLayersByProperty("name", name)[0])
+  $(`#o-legend-${name}`).remove();
+  map.removeLayer(getLayersByProperty('name', name)[0]);
 }
 
 function removeOverlays(overlays) {
   if (overlays) {
-    if (overlays.constructor === Array || overlays instanceof ol.Collection) {
-      overlays.forEach(function(overlay) {
+    if (overlays.constructor === Array || overlays instanceof Collection) {
+      overlays.forEach((overlay) => {
         map.removeOverlay(overlay);
-      })
+      });
     } else {
       map.removeOverlay(overlays);
     }
   } else {
     overlays = map.getOverlays();
-    overlays.forEach(function(overlay) {
-      if (!(overlay.getId() == "geolocation-marker")){
+    overlays.forEach((overlay) => {
+      if (!(overlay.getId() === 'geolocation-marker')) {
         map.removeOverlay(overlay);
       }
     });
   }
 }
 
-function render(el, mapOptions) {
-  pageSettings = mapOptions.pageSettings;
-  pageTemplate.mapClass = "o-map";
+function init(el, mapOptions) {
+  render(el, mapOptions);
 
-  if (pageSettings) {
-    if (pageSettings.footer) {
-      if (pageSettings.footer.hasOwnProperty('img')) {
-        pageTemplate.img = pageSettings.footer.img;
-      }
-      if (pageSettings.footer.hasOwnProperty('text')) {
-        pageTemplate.text = pageSettings.footer.text;
-      }
-      if (pageSettings.footer.hasOwnProperty('url')) {
-        pageTemplate.url = pageSettings.footer.url;
-      }
-      if (pageSettings.footer.hasOwnProperty('urlText')) {
-        pageTemplate.urlText = pageSettings.footer.urlText;
-      }
-    }
-    if (pageSettings.mapGrid) {
-      if (pageSettings.mapGrid.hasOwnProperty('visible') && pageSettings.mapGrid.visible === true) {
-        pageTemplate.mapClass = "o-map o-map-grid";
+  // Read and set projection
+  if ('proj4Defs' in mapOptions && proj4) {
+    project.setProj4(proj4);
+    const proj = mapOptions.proj4Defs;
+
+    // Register proj4 projection definitions
+    for (let i = 0; i < proj.length; i += 1) {
+      proj4.defs(proj[i].code, proj[i].projection);
+      if (Object.prototype.hasOwnProperty.call(proj[i], 'alias')) {
+        proj4.defs(proj[i].alias, proj4.defs(proj[i].code));
       }
     }
   }
+  urlParams = mapOptions.params || {};
+  settings.params = urlParams;
+  settings.map = mapOptions.map;
+  settings.url = mapOptions.url;
+  settings.target = mapOptions.target;
+  settings.baseUrl = mapOptions.baseUrl;
+  settings.breakPoints = mapOptions.breakPoints;
+  settings.capabilitiesURL = mapOptions.capabilitiesURL;
+  settings.config = mapOptions.config;
+  settings.extent = mapOptions.extent || undefined;
+  settings.center = urlParams.center || mapOptions.center;
+  settings.zoom = urlParams.zoom || mapOptions.zoom;
+  settings.tileGrid = mapOptions.tileGrid || {};
+  settings.tileSize = settings.tileGrid.tileSize ? [settings.tileGrid.tileSize, settings.tileGrid.tileSize] : [256, 256];
+  settings.alignBottomLeft = settings.tileGrid.alignBottomLeft;
 
-  $(el).html(template(pageTemplate));
+  if ('proj4Defs' in mapOptions || mapOptions.projectionCode === 'EPSG:3857' || mapOptions.projectionCode === 'EPSG:4326') {
+    // Projection to be used in map
+    settings.projectionCode = mapOptions.projectionCode || undefined;
+    settings.projectionExtent = mapOptions.projectionExtent;
+    settings.projection = new Projection({
+      code: settings.projectionCode,
+      extent: settings.projectionExtent,
+      units: getUnits(settings.projectionCode)
+    });
+    settings.resolutions = mapOptions.resolutions || undefined;
+    settings.tileGrid = maputils.tileGrid(settings);
+  }
+
+  settings.source = mapOptions.source;
+  settings.groups = mapOptions.groups;
+  settings.editLayer = mapOptions.editLayer;
+  settings.styles = mapOptions.styles;
+  settings.clusterOptions = mapOptions.clusterOptions || {};
+  style().init();
+  settings.controls = mapOptions.controls;
+  settings.consoleId = mapOptions.consoleId || 'o-console';
+  settings.featureinfoOptions = mapOptions.featureinfoOptions || {};
+  settings.enableRotation = mapOptions.enableRotation !== false;
+
+  loadMap();
+  settings.layers = createLayers(mapOptions.layers, urlParams.layers);
+  addLayers(settings.layers);
+
+  elQuery(map, {
+    breakPoints: mapOptions.breakPoints,
+    breakPointsPrefix: mapOptions.breakPointsPrefix
+  });
+
+  if (urlParams.feature) {
+    const featureId = urlParams.feature;
+    const layer = getLayer(featureId.split('.')[0]);
+    if (layer) {
+      layer.once('render', () => {
+        let feature;
+        const type = layer.get('type');
+        if (type === 'WFS') {
+          feature = layer.getSource().getFeatureById(featureId);
+        } else {
+          feature = layer.getSource().getFeatureById(featureId.split('.')[1]);
+        }
+
+        if (feature) {
+          const obj = {};
+          obj.feature = feature;
+          obj.title = layer.get('title');
+          obj.content = getattributes(feature, layer);
+          obj.layer = layer;
+
+          const showOverlay = Object.prototype.hasOwnProperty.call(settings.featureinfoOptions, 'overlay') ? settings.featureinfoOptions.overlay : true;
+
+          if (showOverlay) {
+            featureinfo.identify([obj], 'overlay', getcenter(feature.getGeometry()));
+          } else {
+            featureinfo.identify([obj], 'sidebar', getcenter(feature.getGeometry()));
+          }
+          maputils.zoomToExent(feature.getGeometry(), getResolutions().length - 2);
+        }
+      });
+    }
+  }
+  if (urlParams.pin) {
+    settings.featureinfoOptions.savedPin = urlParams.pin;
+  } else if (urlParams.selection) {
+    // This needs further development for proper handling in permalink
+    settings.featureinfoOptions.savedSelection = new Feature({
+      geometry: new geom[urlParams.selection.geometryType](urlParams.selection.coordinates)
+    });
+  }
+  featureinfo.init(settings.featureinfoOptions);
 }
 
-module.exports.init = init;
-module.exports.createLayers = createLayers;
-module.exports.getBaseUrl = getBaseUrl;
-module.exports.getBreakPoints = getBreakPoints;
-module.exports.getExtent = getExtent;
-module.exports.getSettings = getSettings;
-module.exports.getStyleSettings = getStyleSettings;
-module.exports.getMapUrl = getMapUrl;
-module.exports.getMap = getMap;
-module.exports.getLayers = getLayers;
-module.exports.getLayersByProperty = getLayersByProperty;
-module.exports.getLayer = getLayer;
-module.exports.getLayerGroups = getLayerGroups;
-module.exports.getControlNames = getControlNames;
-module.exports.getQueryableLayers = getQueryableLayers;
-module.exports.getSearchableLayers = getSearchableLayers;
-module.exports.getGroup = getGroup;
-module.exports.getGroups = getGroups;
-module.exports.getProjectionCode = getProjectionCode;
-module.exports.getProjection = getProjection;
-module.exports.getMapSource = getMapSource;
-module.exports.getResolutions = getResolutions;
-module.exports.getScale = getScale;
-module.exports.getTarget = getTarget;
-module.exports.getClusterOptions = getClusterOptions;
-module.exports.getTileGrid = getTileGrid;
-module.exports.getTileSize = getTileSize;
-module.exports.autoPan = autoPan;
-module.exports.removeLayer = removeLayer;
-module.exports.removeOverlays = removeOverlays;
-module.exports.checkScale= checkScale;
-module.exports.getMapName = getMapName;
-module.exports.getConsoleId = getConsoleId;
-module.exports.getUrl = getUrl;
-module.exports.getConfig = getConfig;
+export default {
+  init,
+  createLayers,
+  getBaseUrl,
+  getBreakPoints,
+  getConfig,
+  getExtent,
+  getSettings,
+  getStyleSettings,
+  getMapUrl,
+  getMap,
+  getLayers,
+  getLayersByProperty,
+  getLayer,
+  getLayerGroups,
+  getControlNames,
+  getQueryableLayers,
+  getSearchableLayers,
+  getGroup,
+  getGroups,
+  getProjectionCode,
+  getProjection,
+  getMapSource,
+  getResolutions,
+  getScale,
+  getTarget,
+  getClusterOptions,
+  getTileGrid,
+  getTileSize,
+  removeLayer,
+  removeOverlays,
+  checkScale,
+  getMapName,
+  getConsoleId,
+  getUrl
+};
