@@ -1,148 +1,137 @@
-"use strict";
+import Point from 'ol/geom/point';
+import Circle from 'ol/style/circle';
+import Fill from 'ol/style/fill';
+import Icon from 'ol/style/icon';
+import Stroke from 'ol/style/stroke';
+import Style from 'ol/style/style';
+import Text from 'ol/style/text';
+import $ from 'jquery';
+import viewer from './viewer';
+import validateurl from './utils/validateurl';
+import stylefunctions from './style/stylefunctions';
+import replacer from '../src/utils/replacer';
+import maputils from './maputils';
 
-var ol = require('openlayers');
-var $ = require('jquery');
-var Viewer = require('./viewer');
-var validateUrl = require('./utils/validateurl');
-var styleFunctions = require('./style/stylefunctions');
-var replacer = require('../src/utils/replacer');
-var mapUtils = require('./maputils');
+let baseUrl;
 
-var baseUrl;
+const white = [255, 255, 255, 1];
+const blue = [0, 153, 255, 1];
+const width = 3;
 
-var white = [255, 255, 255, 1];
-var blue = [0, 153, 255, 1];
-var width = 3;
-
-//default edit style options
-var editStyleOptions = {
-  'Point': [{
-    "circle": {
-      "radius": 1,
-      "stroke": {
-        "color": blue,
-        "width": 0
+// default edit style options
+const editStyleOptions = {
+  Point: [{
+    circle: {
+      radius: 1,
+      stroke: {
+        color: blue,
+        width: 0
       },
-      "fill": {
-        "color": blue
+      fill: {
+        color: blue
       }
     }
   }],
-  'LineString': [{
-      'stroke': {
-        color: white,
-        width: width + 2
-      }
-    },
-    {
-      'stroke': {
-        color: blue,
-        width: width
-      }
+  LineString: [{
+    stroke: {
+      color: white,
+      width: width + 2
     }
+  },
+  {
+    stroke: {
+      color: blue,
+      width
+    }
+  }
   ],
-  'Polygon': [{
-      'stroke': {
-        color: white,
-        width: width + 2
-      }
-    },
-    {
-      'stroke': {
-        color: blue,
-        width: width
-      }
+  Polygon: [{
+    stroke: {
+      color: white,
+      width: width + 2
     }
+  },
+  {
+    stroke: {
+      color: blue,
+      width
+    }
+  }
   ]
 };
 
-module.exports = function() {
-
-  return {
-    init: Init,
-    createStyleOptions: createStyleOptions,
-    createStyleList: createStyleList,
-    createStyleRule: createStyleRule,
-    createStyle: createStyle,
-    styleFunction: styleFunction,
-    createEditStyle: createEditStyle,
-    createGeometryStyle: createGeometryStyle
-  };
-};
-
-function Init() {
-  baseUrl = Viewer.getBaseUrl();
-}
-
 function createStyleOptions(styleParams) {
-  var styleOptions = {};
-  if (styleParams.hasOwnProperty('geometry')) {
+  const styleOptions = {};
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'geometry')) {
     switch (styleParams.geometry) {
       case 'centerPoint':
-        styleOptions.geometry = function(feature) {
-          var coordinates = mapUtils.getCenter(feature.getGeometry());
-          return new ol.geom.Point(coordinates);
-        }
+        styleOptions.geometry = function centerPoint(feature) {
+          const coordinates = maputils.getCenter(feature.getGeometry());
+          return new Point(coordinates);
+        };
         break;
       case 'endPoint':
-        styleOptions.geometry = function(feature) {
-          var coordinates = feature.getGeometry().getLastCoordinate();
-          return new ol.geom.Point(coordinates);
-        }
+        styleOptions.geometry = function endPoint(feature) {
+          const coordinates = feature.getGeometry().getLastCoordinate();
+          return new Point(coordinates);
+        };
         break;
+      default:
+      {
+        break;
+      }
     }
   }
-  if (styleParams.hasOwnProperty('zIndex')) {
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'zIndex')) {
     styleOptions.zIndex = styleParams.zIndex;
   }
-  if (styleParams.hasOwnProperty('fill')) {
-    styleOptions.fill = new ol.style.Fill(styleParams.fill);
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'fill')) {
+    styleOptions.fill = new Fill(styleParams.fill);
   }
-  if (styleParams.hasOwnProperty('stroke')) {
-    styleOptions.stroke = new ol.style.Stroke(styleParams.stroke);
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'stroke')) {
+    styleOptions.stroke = new Stroke(styleParams.stroke);
   }
-  if (styleParams.hasOwnProperty('text')) {
-    styleOptions.text = new ol.style.Text(styleParams.text);
-    if (styleParams.text.hasOwnProperty('fill')) {
-      styleOptions.text.setFill(new ol.style.Fill(styleParams.text.fill));
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'text')) {
+    styleOptions.text = new Text(styleParams.text);
+    if (Object.prototype.hasOwnProperty.call(styleParams.text, 'fill')) {
+      styleOptions.text.setFill(new Fill(styleParams.text.fill));
     }
-    if (styleParams.text.hasOwnProperty('stroke')) {
-      styleOptions.text.setStroke(new ol.style.Stroke(styleParams.text.stroke));
+    if (Object.prototype.hasOwnProperty.call(styleParams.text, 'stroke')) {
+      styleOptions.text.setStroke(new Stroke(styleParams.text.stroke));
     }
   }
-  if (styleParams.hasOwnProperty('icon')) {
-    if (styleParams.icon.hasOwnProperty('src')) {
-      styleParams.icon.src = validateUrl(styleParams.icon.src, baseUrl);
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'icon')) {
+    const styleIcon = styleParams.icon;
+    if (Object.prototype.hasOwnProperty.call(styleIcon, 'src')) {
+      styleIcon.src = validateurl(styleIcon.src, baseUrl);
     }
-    styleOptions.image = new ol.style.Icon(styleParams.icon);
+    styleOptions.image = new Icon(styleIcon);
   }
-  if (styleParams.hasOwnProperty('circle')) {
-    styleOptions.image = new ol.style.Circle({
+  if (Object.prototype.hasOwnProperty.call(styleParams, 'circle')) {
+    styleOptions.image = new Circle({
       radius: styleParams.circle.radius,
-      fill: new ol.style.Fill(styleParams.circle.fill) || undefined,
-      stroke: new ol.style.Stroke(styleParams.circle.stroke) || undefined
+      fill: new Fill(styleParams.circle.fill) || undefined,
+      stroke: new Stroke(styleParams.circle.stroke) || undefined
     });
   }
   return styleOptions;
 }
 
 function createStyleList(styleOptions) {
-  var styleList = [];
-  //Create style for each rule
-  for (var i = 0; i < styleOptions.length; i++) {
-    var styleRule = [];
-    var styleOption;
-    //Check if rule is array, ie multiple styles for the rule
+  const styleList = [];
+  // Create style for each rule
+  for (let i = 0; i < styleOptions.length; i += 1) {
+    let styleRule = [];
+    let styleOption;
+    // Check if rule is array, ie multiple styles for the rule
     if (styleOptions[i].constructor === Array) {
-      for (var j = 0; j < styleOptions[i].length; j++) {
+      for (let j = 0; j < styleOptions[i].length; j += 1) {
         styleOption = createStyleOptions(styleOptions[i][j]);
-        styleRule.push(new ol.style.Style(styleOption));
+        styleRule.push(new Style(styleOption));
       }
-    }
-    //If single style for rule
-    else {
+    } else {
       styleOption = createStyleOptions(styleOptions[i]);
-      styleRule = [new ol.style.Style(styleOption)];
+      styleRule = [new Style(styleOption)];
     }
 
     styleList.push(styleRule);
@@ -150,110 +139,39 @@ function createStyleList(styleOptions) {
   return styleList;
 }
 
-function createStyle(styleName, clusterStyleName) {
-  var styleSettings = Viewer.getStyleSettings()[styleName];
-  if ($.isEmptyObject(styleSettings)) {
-    alert('Style ' + styleName + ' is not defined');
-  }
-  if (styleSettings[0][0].hasOwnProperty("custom")) {
-    var style = styleFunctions(styleSettings[0][0].custom, styleSettings[0][0].params);
-    return style;
-  } else {
-    var clusterStyleSettings = Viewer.getStyleSettings()[clusterStyleName];
-    var style = (function() {
-      //Create style for each rule
-      var styleList = createStyleList(styleSettings);
-      if (clusterStyleSettings) {
-        var clusterStyleList = createStyleList(clusterStyleSettings);
-        return styleFunction(styleSettings, styleList, clusterStyleSettings, clusterStyleList);
-      } else {
-        return styleFunction(styleSettings, styleList);
-      }
-
-    })()
-    return style;
-  }
-}
-
-function createStyleRule(options) {
-  var styleRule = [],
-    styleOption;
-  if (options.constructor === Array) {
-    for (var i = 0; i < options.length; i++) {
-      var styleOption = createStyleOptions(options[i]);
-      styleRule.push(new ol.style.Style(styleOption));
-    }
-  }
-  //If single style for rule
-  else {
-    styleOption = createStyleOptions(options);
-    styleRule = [new ol.style.Style(styleOption)];
-  }
-  return styleRule;
-}
-
-function styleFunction(styleSettings, styleList, clusterStyleSettings, clusterStyleList) {
-  var s = styleSettings;
-  var resolutions = Viewer.getResolutions();
-  var fn = function(feature, resolution) {
-    var scale = Viewer.getScale(resolution);
-    var styleL;
-    //If size is larger than, it is a cluster
-    var size = clusterStyleList ? feature.get('features').length : 1;
-    if (size > 1 && resolution != resolutions[resolutions.length + 1]) {
-      styleL = checkOptions(feature, scale, clusterStyleSettings, clusterStyleList, size.toString());
-      // clusterStyleList[0].setText(size);
-    } else {
-      styleL = checkOptions(feature, scale, styleSettings, styleList);
-    }
-    return styleL;
-  }
-  return fn;
-}
-
-function createEditStyle() {
-  return createGeometryStyle(editStyleOptions);
-}
-
-function createGeometryStyle(geometryStyleOptions) {
-  return {
-    'Point': createStyleRule(geometryStyleOptions.Point),
-    'MultiPoint': createStyleRule(geometryStyleOptions.Point),
-    'LineString': createStyleRule(geometryStyleOptions.LineString),
-    'MultiLineString': createStyleRule(geometryStyleOptions.LineString),
-    'Polygon': createStyleRule(geometryStyleOptions.Polygon),
-    'MultiPolygon': createStyleRule(geometryStyleOptions.Polygon)
-  };
-}
-
 function checkOptions(feature, scale, styleSettings, styleList, size) {
-  var s = styleSettings;
-  for (var j = 0; j < s.length; j++) {
-    var styleL;
-    if (Viewer.checkScale(scale, s[j][0].maxScale, s[j][0].minScale)) {
-      s[j].some(function(element, index, array) {
-        if (element.hasOwnProperty('text') && size) {
+  const s = styleSettings;
+  for (let j = 0; j < s.length; j += 1) {
+    let styleL;
+    if (viewer.checkScale(scale, s[j][0].maxScale, s[j][0].minScale)) {
+      s[j].some((element, index) => {
+        if (Object.prototype.hasOwnProperty.call(element, 'text') && size) {
           styleList[j][index].getText().setText(size);
-        } else if (element.hasOwnProperty('text')) {
+        } else if (Object.prototype.hasOwnProperty.call(element, 'text')) {
           styleList[j][index].getText().setText(replacer.replace(element.text.text, feature.getProperties()));
         }
+        return null;
       });
-      if (s[j][0].hasOwnProperty('filter')) {
-        //find attribute vale between [] defined in styles
-        var featAttr, expr, regexExpr, regexFilter, featMatch;
-        var matches = s[j][0].filter.match(/\[(.*?)\]/);
+      if (Object.prototype.hasOwnProperty.call(s[j][0], 'filter')) {
+        let expr;
+        // find attribute vale between [] defined in styles
+
+        let regexExpr;
+        let regexFilter;
+        let featMatch;
+
+        const matches = s[j][0].filter.match(/\[(.*?)\]/);
         if (matches) {
+          let first = feature;
           if (feature.get('features')) {
-            feature = feature.get('features')[0];
+            first = feature.get('features')[0];
           }
-          featAttr = matches[1];
+          const featAttr = matches[1];
           expr = s[j][0].filter.split(']')[1];
-          featMatch = feature.get(featAttr);
+          featMatch = first.get(featAttr);
           regexFilter = s[j][0].filter.match(/\/(.*)\/([a-zA-Z]+)?/);
-
-          expr = typeof featMatch == 'number' ? featMatch + expr : '"' + featMatch + '"' + expr;
+          expr = typeof featMatch === 'number' ? featMatch + expr : `"${featMatch}"${expr}`;
         }
-
         if (regexFilter) {
           regexExpr = new RegExp(regexFilter[1], regexFilter[2]);
 
@@ -261,11 +179,10 @@ function checkOptions(feature, scale, styleSettings, styleList, size) {
             styleL = styleList[j];
             return styleL;
           }
-        } else {
-          if (eval(expr)) {
-            styleL = styleList[j];
-            return styleL;
-          }
+        }
+        if (eval(expr)) {
+          styleL = styleList[j];
+          return styleL;
         }
       } else {
         styleL = styleList[j];
@@ -273,4 +190,92 @@ function checkOptions(feature, scale, styleSettings, styleList, size) {
       }
     }
   }
+  return null;
+}
+
+function styleFunction(styleSettings, styleList, clusterStyleSettings, clusterStyleList) {
+  const resolutions = viewer.getResolutions();
+  const fn = function fn(feature, resolution) {
+    const scale = viewer.getScale(resolution);
+    let styleL;
+    // If size is larger than, it is a cluster
+    const size = clusterStyleList ? feature.get('features').length : 1;
+    if (size > 1 && resolution !== resolutions[resolutions.length + 1]) {
+      styleL = checkOptions(feature, scale, clusterStyleSettings, clusterStyleList, size.toString());
+      // clusterStyleList[0].setText(size);
+    } else {
+      styleL = checkOptions(feature, scale, styleSettings, styleList);
+    }
+    return styleL;
+  };
+  return fn;
+}
+
+function createStyle(styleName, clusterStyleName) {
+  const styleSettings = viewer.getStyleSettings()[styleName];
+  if ($.isEmptyObject(styleSettings)) {
+    alert(`Style ${styleName} is not defined`);
+  }
+  if (Object.prototype.hasOwnProperty.call(styleSettings[0][0], 'custom')) {
+    const style = stylefunctions(styleSettings[0][0].custom, styleSettings[0][0].params);
+    return style;
+  }
+  const clusterStyleSettings = viewer.getStyleSettings()[clusterStyleName];
+  const style = (function style() {
+    // Create style for each rule
+    const styleList = createStyleList(styleSettings);
+    if (clusterStyleSettings) {
+      const clusterStyleList = createStyleList(clusterStyleSettings);
+      return styleFunction(styleSettings, styleList, clusterStyleSettings, clusterStyleList);
+    }
+    return styleFunction(styleSettings, styleList);
+  }());
+  return style;
+}
+
+function createStyleRule(options) {
+  let styleRule = [];
+  let styleOption;
+  if (options.constructor === Array) {
+    for (let i = 0; i < options.length; i += 1) {
+      styleOption = createStyleOptions(options[i]);
+      styleRule.push(new Style(styleOption));
+    }
+  } else {
+    styleOption = createStyleOptions(options);
+    styleRule = [new Style(styleOption)];
+  }
+  return styleRule;
+}
+
+function createGeometryStyle(geometryStyleOptions) {
+  return {
+    Point: createStyleRule(geometryStyleOptions.Point),
+    MultiPoint: createStyleRule(geometryStyleOptions.Point),
+    LineString: createStyleRule(geometryStyleOptions.LineString),
+    MultiLineString: createStyleRule(geometryStyleOptions.LineString),
+    Polygon: createStyleRule(geometryStyleOptions.Polygon),
+    MultiPolygon: createStyleRule(geometryStyleOptions.Polygon)
+  };
+}
+
+function createEditStyle() {
+  return createGeometryStyle(editStyleOptions);
+}
+
+function Init() {
+  baseUrl = viewer.getBaseUrl();
+}
+
+export default function () {
+  return {
+    init: Init,
+    createStyleOptions,
+    createStyleList,
+    createStyleRule,
+    createStyle,
+    styleFunction,
+    createEditStyle,
+    createGeometryStyle
+  };
 }
